feat: autoconfig package adjustment，Distinguish springboot 2、3
