+++ dify/spring-boot-starter/dify-spring-boot-starter-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,6 +1,6 @@
-io.github.guoshiqiufeng.dify.autoconfigure.DifyPropertiesAutoConfiguration
-io.github.guoshiqiufeng.dify.autoconfigure.DifyChatAutoConfiguration
-io.github.guoshiqiufeng.dify.autoconfigure.DifyDatasetAutoConfiguration
-io.github.guoshiqiufeng.dify.autoconfigure.DifyServerRedisTokenAutoConfiguration
-io.github.guoshiqiufeng.dify.autoconfigure.DifyServerAutoConfiguration
-io.github.guoshiqiufeng.dify.autoconfigure.DifyWorkflowAutoConfiguration
+io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyPropertiesAutoConfiguration
+io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyChatAutoConfiguration
+io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyDatasetAutoConfiguration
+io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyServerRedisTokenAutoConfiguration
+io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyServerAutoConfiguration
+io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyWorkflowAutoConfiguration
