+++ dify/spring-boot-starter/dify-spring-boot-starter-autoconfigure/src/test/java/io/github/guoshiqiufeng/dify/springboot/autoconfigure/DifyPropertiesAutoConfigurationTest.java
@@ -15,8 +15,6 @@
  */
 package io.github.guoshiqiufeng.dify.springboot.autoconfigure;
 
-import io.github.guoshiqiufeng.dify.autoconfigure.DifyConnectionDetails;
-import io.github.guoshiqiufeng.dify.autoconfigure.DifyPropertiesAutoConfiguration;
 import io.github.guoshiqiufeng.dify.core.config.DifyProperties;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
