+++ dify/spring-boot-starter/dify-spring-boot-starter-autoconfigure/src/test/java/io/github/guoshiqiufeng/dify/springboot/autoconfigure/DifyWorkflowAutoConfigurationTest.java
@@ -15,8 +15,6 @@
  */
 package io.github.guoshiqiufeng.dify.springboot.autoconfigure;
 
-import io.github.guoshiqiufeng.dify.autoconfigure.DifyPropertiesAutoConfiguration;
-import io.github.guoshiqiufeng.dify.autoconfigure.DifyWorkflowAutoConfiguration;
 import io.github.guoshiqiufeng.dify.client.spring6.workflow.DifyWorkflowDefaultClient;
 import io.github.guoshiqiufeng.dify.core.config.DifyProperties;
 import io.github.guoshiqiufeng.dify.workflow.DifyWorkflow;
