+++ dify/spring-boot2-starter/dify-spring-boot2-starter-autoconfigure/src/main/java/io/github/guoshiqiufeng/dify/springboot2/autoconfigure/DifyDatasetAutoConfiguration.java
@@ -13,19 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.guoshiqiufeng.dify.autoconfigure;
+package io.github.guoshiqiufeng.dify.springboot2.autoconfigure;
 
-import io.github.guoshiqiufeng.dify.chat.DifyChat;
-import io.github.guoshiqiufeng.dify.chat.client.DifyChatClient;
-import io.github.guoshiqiufeng.dify.chat.impl.DifyChatClientImpl;
-import io.github.guoshiqiufeng.dify.client.spring5.chat.DifyChatDefaultClient;
+import io.github.guoshiqiufeng.dify.client.spring5.dataset.DifyDatasetDefaultClient;
 import io.github.guoshiqiufeng.dify.core.config.DifyProperties;
+import io.github.guoshiqiufeng.dify.dataset.DifyDataset;
+import io.github.guoshiqiufeng.dify.dataset.client.DifyDatasetClient;
+import io.github.guoshiqiufeng.dify.dataset.impl.DifyDatasetClientImpl;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpHeaders;
 import org.springframework.web.reactive.function.client.WebClient;
 
 /**
@@ -35,22 +36,25 @@ import org.springframework.web.reactive.function.client.WebClient;
  */
 @Slf4j
 @Configuration
-@ConditionalOnClass({DifyChatClient.class})
-public class DifyChatAutoConfiguration {
+@ConditionalOnClass({DifyDatasetClient.class})
+public class DifyDatasetAutoConfiguration {
 
     @Bean
-    @ConditionalOnMissingBean(DifyChatClient.class)
-    public DifyChatClient difyChatClient(DifyProperties properties,
-                                         ObjectProvider<WebClient.Builder> webClientBuilderProvider) {
-        return new DifyChatDefaultClient(properties.getUrl(),
+    @ConditionalOnMissingBean(DifyDatasetClient.class)
+    public DifyDatasetClient difyDatasetClient(DifyProperties properties,
+                                               ObjectProvider<WebClient.Builder> webClientBuilderProvider) {
+        String apikey = "Bearer " + properties.getDataset().getApiKey();
+        return new DifyDatasetDefaultClient(properties.getUrl(),
                 properties.getClientConfig(),
-                webClientBuilderProvider.getIfAvailable(WebClient::builder));
+                webClientBuilderProvider.getIfAvailable(WebClient::builder)
+                        .defaultHeader(HttpHeaders.AUTHORIZATION, apikey));
     }
 
     @Bean
-    @ConditionalOnMissingBean({DifyChat.class})
-    public DifyChatClientImpl difyChatHandler(DifyChatClient difyChatClient) {
-        return new DifyChatClientImpl(difyChatClient);
+    @ConditionalOnMissingBean({DifyDataset.class})
+    public DifyDatasetClientImpl difyDataset(DifyDatasetClient difyDatasetClient) {
+        return new DifyDatasetClientImpl(difyDatasetClient);
     }
+
 
 }
