+++ dify/spring-boot2-starter/dify-spring-boot2-starter-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,6 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  io.github.guoshiqiufeng.dify.autoconfigure.DifyPropertiesAutoConfiguration,\
-  io.github.guoshiqiufeng.dify.autoconfigure.DifyChatAutoConfiguration,\
-  io.github.guoshiqiufeng.dify.autoconfigure.DifyDatasetAutoConfiguration,\
-  io.github.guoshiqiufeng.dify.autoconfigure.DifyServerAutoConfiguration,\
-  io.github.guoshiqiufeng.dify.autoconfigure.DifyWorkflowAutoConfiguration
+  io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyPropertiesAutoConfiguration,\
+  io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyChatAutoConfiguration,\
+  io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyDatasetAutoConfiguration,\
+  io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyServerAutoConfiguration,\
+  io.github.guoshiqiufeng.dify.springboot2.autoconfigure.DifyWorkflowAutoConfiguration
