+++ dify/spring-boot2-starter/dify-spring-boot2-starter-autoconfigure/src/test/java/io/github/guoshiqiufeng/dify/springboot2/autoconfigure/DifyServerRedisTokenAutoConfigurationTest.java
@@ -15,8 +15,6 @@
  */
 package io.github.guoshiqiufeng.dify.springboot2.autoconfigure;
 
-import io.github.guoshiqiufeng.dify.autoconfigure.DifyPropertiesAutoConfiguration;
-import io.github.guoshiqiufeng.dify.autoconfigure.DifyServerRedisTokenAutoConfiguration;
 import io.github.guoshiqiufeng.dify.server.client.BaseDifyServerToken;
 import io.github.guoshiqiufeng.dify.server.client.DifyServerClient;
 import io.github.guoshiqiufeng.dify.server.client.DifyServerTokenDefault;
